+++ topic-operator/src/test/java/io/strimzi/operator/topic/TopicControllerIT.java
@@ -1231,10 +1231,11 @@ class TopicControllerIT {
                 TopicControllerIT.class.getSimpleName(), 
                 fullReconciliationIntervalMs,
                 false, "", "", "", "", "",
-                false, "", "", "", "",
+                false, "", "", "", "", "",
                 useFinalizer,
                 100, 100, 10, false,
-                false, false, "", 9090, false, false, "", "", "");
+                false, false, "", 9090, false, false, "", "", "",
+                "all", false);
     }
 
     @ParameterizedTest
@@ -1984,10 +1985,11 @@ class TopicControllerIT {
         var config = new TopicOperatorConfig(ns, Labels.fromMap(SELECTOR),
                 kafkaCluster.getBootstrapServers(), TopicControllerIT.class.getSimpleName(), 10_000,
                 false, "", "", "", "", "",
-                false, "", "", "", "",
+                false, "", "", "", "", "",
                 true,
                 1, 100, 5_0000, false,
-                false, false, "", 9090, false, false, "", "", "");
+                false, false, "", 9090, false, false, "", "", "",
+                "all", false);
 
         maybeStartOperator(config);